requirements.txt
# Core API Framework
fastapi==0.110.0
uvicorn[standard]==0.29.0

# MongoDB (Using Models via MongoEngine)
mongoengine==0.27.0
pymongo==4.6.1

# Environment Variables
python-dotenv==1.0.1

# LLM (Feature-flag controlled)
openai==1.12.0

# Validation (FastAPI dependency)
pydantic==2.6.4

